Starting on data analysis; stopping work on poll
PoliSci class has determined that we will use Google Forms for our
survey, instead of a custom polling program, so I'm stopping work on the
polling program. (Kind of a relief since I was struggling with TKinter.)
However, we do need a tool to analyze all the data we get out. This
data_analysis.py is that tool. It will let us filter the responses for
each question by group to get responses by percentage and compare across
groups.