Additional GUI items, GUI information retrieval
"Question" dropdown where you can select the question you're interested
in; "Submit Query" button that will be bound to the data retrieval
command.
Data retrieval now automatically gets the GUI demographics selections to
cull voters. Need to do similar automation for the question and answer
selections. (Answer selections will need to be bound to question
selections because the answers available in the dropdown depend on the
question.)