Half-baked func to analyze stance by demographic; fixed stance calc
Wrote function that iterates through demographics and finds the average
stance; doesn't work yet but is close.
Also, changed the individual stance calculation to put the index on the
scale from 1 - 5 like the questions.

PROBLEM: when the data is retrieved and printed during a query, the
responses are printed in random order as a result of the random behavior
of dict keys. Will need to go through and change dicts to OrderedDicts
in order to fix this :|